feat: load minimal code if no SNS Adaptor started
