+++ packages/maskbook/src/content-script.ts
@@ -1,3 +1,6 @@
 import './extension/content-script/hmr'
-import './setup.ui'
-import './extension/content-script/index'
+import { status } from './setup.ui'
+
+status.then((loaded) => {
+    loaded && import('./extension/content-script/tasks')
+})
