+++ packages/maskbook/src/social-network-adaptor/index.ts
@@ -1,6 +1,3 @@
 import './facebook.com'
 import './twitter.com'
 import './instagram.com'
-
-import './options-page'
-import './popup-page'
