+++ packages/maskbook/src/social-network/index.ts
@@ -1,4 +1,5 @@
 export * from './types'
 export * from './utils'
 export * from './ui'
+export * from './define'
 export * from './worker'
