+++ packages/maskbook/src/utils/native-rpc/Web.ts
@@ -1,6 +1,6 @@
 import type { WebviewAPIs } from './types'
 import Services from '../../extension/service'
-import { definedSocialNetworkWorkers } from '../../social-network/worker'
+import { definedSocialNetworkWorkers } from '../../social-network/define'
 import { launchPageSettings } from '../../settings/settings'
 import stringify from 'json-stable-stringify'
 import { unreachable } from '../utils'
